v1

+++ traceroute/prettyprint.h
@@ -1,4 +1,4 @@
 #ifndef PRETTYPRINT_H
 # define PRETTYPRINT_H
-void pretty_print(int received, char sender_ip_str[3][20], int time, int num);
+void pretty_print(int received, char sender_ip_str[3][20], double time, int num);
 #endif
